+++ decidim-accountability/spec/commands/admin/create_result_spec.rb
@@ -125,6 +125,49 @@ module Decidim::Accountability
 
         expect(linked_meetings).to eq [meeting]
       end
+
+      it "notifies the process followers" do
+        follower = create(:user, organization: organization)
+        create(:follow, followable: proposals.first, user: follower)
+
+        expect(Decidim::EventsManager)
+          .to receive(:publish)
+          .with(
+            event: "decidim.events.accountability.proposal_linked",
+            event_class: Decidim::Accountability::ProposalLinkedEvent,
+            resource: kind_of(Result),
+            recipient_ids: [proposals.first.author.id, follower.id],
+            extra: {
+              proposal_id: proposals.first.id
+            }
+          )
+
+        expect(Decidim::EventsManager)
+          .to receive(:publish)
+          .with(
+            event: "decidim.events.accountability.proposal_linked",
+            event_class: Decidim::Accountability::ProposalLinkedEvent,
+            resource: kind_of(Result),
+            recipient_ids: [proposals.second.author.id],
+            extra: {
+              proposal_id: proposals.second.id
+            }
+          )
+
+        expect(Decidim::EventsManager)
+          .to receive(:publish)
+          .with(
+            event: "decidim.events.accountability.proposal_linked",
+            event_class: Decidim::Accountability::ProposalLinkedEvent,
+            resource: kind_of(Result),
+            recipient_ids: [proposals.third.author.id],
+            extra: {
+              proposal_id: proposals.third.id
+            }
+          )
+
+        subject.call
+      end
     end
   end
 end